fix(llmcontext): key archive by memory seq; cron fingerprint collapse
Root cause: the archive used a private per-session counter that diverged from
the memory/conversation seq as an agent aged. archiveWindow() returned those
archive-space bounds, so StripOutOfRangeSeqRefs deleted every summary ref whose
seq exceeded the (much lower) archive max — gutting all recent state from the
compaction summary on long-lived agents and breaking get_session_messages.

Archive by the memory seq instead, so one sequence space spans the summarizer,
the strip, the archive, and seq-based retrieval. Widen the whole seq path to
int64 to support agents running for years without reset.

Also:
- Summarizer prompts now complement the always-present system prompt, biasing
  toward transient in-flight state instead of re-deriving standing rules.
- Cron jobs get a content-addressed fingerprint (sha256 hex[:8] over schedule +
  payload) injected as a leading [<fp>] marker on each fired message. Repeated
  cron no-op runs are programmatically collapsed (summarizer input AND retained
  tail) into one counted anchor noting the [first-last] seq range; collapsed
  originals stay in the archive, retrievable by seq. A fire whose reply issued
  tool calls (an action) or differs/long-form breaks the run and is preserved.
- Wire WithCompressModel so the summary's "Generated by <model>" stamp populates.

+++ pkg/agent/loop_session_tools.go
@@ -31,8 +31,8 @@ func buildSessionInfo(al *AgentLoop, agent *AgentInstance, sessionKey string) (*
 	rawSummary := agent.Sessions.GetSummary(sessionKey)
 	if rawSummary != "" {
 		var sv struct {
-			CoveredSeqStart int `json:"covered_seq_start"`
-			CoveredSeqEnd   int `json:"covered_seq_end"`
+			CoveredSeqStart int64 `json:"covered_seq_start"`
+			CoveredSeqEnd   int64 `json:"covered_seq_end"`
 		}
 		if err := json.Unmarshal([]byte(rawSummary), &sv); err == nil && sv.CoveredSeqStart > 0 {
 			info.SummaryCovers = &tools.SummaryCoverage{
